requirements.txt
mysqlclient==1.3.10
python-mimeparse==1.6.0
python-dateutil
defusedxml
lxml
pyyaml
django-tastypie==0.13.3
lxml==3.8.0
mysqlclient==1.3.10
python-dateutil==2.6.0
python-mimeparse==1.6.0
PyYAML==3.12
six==1.10.0
django